Replace 4 counter track components with a single generic TrackCounter
  Collapse the separate Memory, Power, ProcessCPU, and Bandwidth track
  implementations into a single TrackCounter component that renders any
  counter type using CounterDisplayConfig from PR #5912.

  The LocalTrack union type is simplified from 8 variants to 5 by
  replacing 'memory', 'power', 'process-cpu', and 'bandwidth' with a
  single 'counter' type. The component branches on display.graphType
  for canvas drawing (accumulated vs rate) and on display.unit for
  tooltip rendering (bytes, pWh, percent, etc.).

  Track index ordering (for URL backward compatibility) is handled by
  a category-based mapping function. Display ordering uses the new
  display.sortWeight field.

  Part of #5752.
