--- src/components/timeline/TrackBandwidth.css
@@ -1,25 +1,0 @@
-/* This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
-
-.timelineTrackBandwidthGraph {
-  position: relative;
-  width: 100%;
-  height: var(--graph-height);
-}
-
-.timelineTrackBandwidthCanvas {
-  position: absolute;
-  width: 100%;
-  height: 100%;
-}
-
-.timelineTrackBandwidthGraphDot {
-  position: absolute;
-  width: 6px;
-  height: 6px;
-  border-radius: 3px;
-  margin-top: -3px;
-  margin-left: -3px;
-  pointer-events: none;
-}
