--- src/components/timeline/TrackMemory.css
@@ -1,41 +1,0 @@
-/* This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
-
-.timelineTrackMemoryGraph {
-  position: relative;
-  width: 100%;
-  height: var(--graph-height);
-}
-
-.timelineTrackMemoryCanvas {
-  position: absolute;
-  width: 100%;
-  height: 100%;
-}
-
-.timelineTrackMemoryGraphDot {
-  position: absolute;
-  width: 6px;
-  height: 6px;
-  border-radius: 3px;
-  margin-top: -3px;
-  margin-left: -3px;
-  pointer-events: none;
-}
-
-.timelineTrackMemoryTooltipLine {
-  white-space: nowrap;
-}
-
-.timelineTrackMemoryTooltipNumber {
-  display: inline-block;
-  min-width: 60px;
-  color: var(--tooltip-number-foreground-color);
-  font-weight: bold;
-}
-
-.timelineMarkersMemory {
-  height: var(--markers-height, 15px);
-  opacity: 1;
-}
