--- src/components/timeline/TrackProcessCPU.css
@@ -1,39 +1,0 @@
-/* This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
-
-.timelineTrackProcessCPUGraph {
-  position: relative;
-  width: 100%;
-  height: var(--graph-height);
-}
-
-.timelineTrackProcessCPUCanvas {
-  position: absolute;
-  width: 100%;
-  height: 100%;
-}
-
-.timelineTrackProcessCPUGraphDot {
-  --internal-background-color: var(--grey-50);
-
-  position: absolute;
-  width: 6px;
-  height: 6px;
-  border-radius: 3px;
-  margin-top: -3px;
-  margin-left: -3px;
-  background-color: var(--internal-background-color);
-  pointer-events: none;
-}
-
-.timelineTrackProcessCPUTooltipLine {
-  white-space: nowrap;
-}
-
-.timelineTrackProcessCPUTooltipNumber {
-  display: inline-block;
-  min-width: 60px;
-  color: var(--tooltip-number-foreground-color);
-  font-weight: bold;
-}
