+++ src/profile-logic/processed-profile-versioning.ts
@@ -3092,7 +3092,7 @@ const _upgraders: {
             unit: 'percent',
             color: 'grey',
             markerSchemaLocation: null,
-            sortWeight: 40,
+            sortWeight: 70,
             label: 'Process CPU',
           };
         } else {
@@ -3101,7 +3101,7 @@ const _upgraders: {
             unit: '',
             color: 'grey',
             markerSchemaLocation: null,
-            sortWeight: 50,
+            sortWeight: 35,
             label: name,
           };
         }
